// src/types/Settings.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AlertSettings } from "./AlertSettings";
import type { DataSourceSettings } from "./DataSourceSettings";
import type { ScoreWeights } from "./ScoreWeights";

export type Settings = { score_weights: ScoreWeights, alerts: AlertSettings, data_sources: DataSourceSettings, };